requirements.txt
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.8.0
asgiref==3.8.1
beam-client==0.2.128
beautifulsoup4==4.12.3
beta9==0.1.142
betterproto-beta9==2.0.0b7
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contractions==0.1.73
croniter==3.0.4
fastapi==0.110.3
filelock==3.16.1
fsspec==2024.12.0
grpcio==1.67.0
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
huggingface-hub==0.27.1
hyperframe==6.0.1
idna==3.10
isort==5.13.2
Jinja2==3.1.5
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
nltk==3.9.1
numpy==2.2.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==4.25.5
pyahocorasick==2.1.0
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.37.2
textsearch==0.0.24
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.48.0
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.29.0
watchdog==4.0.2
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
